Update on "[monarch] ndslice: introduce views"
This adds `View` to the ndslice module.

A view is a collection of ranks, projected into an extent. It is represented as a set of labels and a slice. (Subject to extension later.)

Currently, views can be constructed by subsetting an existing view or an extent. Specifically, we support specifying ranges for each dimension.

Differential Revision: [D79372445](https://our.internmc.facebook.com/intern/diff/D79372445/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D79372445/)!

[ghstack-poisoned]