To fix the shutdown errors, we need to stop the tokio loop...
I am pretty sure our finalization errors are some combo of:

1. schedule some work that will hold or take GIL on a thread owned by tokio. This includes anything that holds a reference to a PyObject, because Drop on PyObject will take the GIL.
2. call Py_Finalize from the main thread
3. Python unloads
4. tokio thread tries to use GIL and crashes.

We also start at least one std::thread which runs python that will encounter this problem and not be on the tokio loop.

To avoid this, we need to shutdown the tokio loop (i.e. allow all tasks to reach an await, and then cancel everything).

However, pyo3_async_runtime makes this nearly impossible because it requires a 'static lifetime reference to the loop, so it cannot be shutdown.

This diff shows how to call shutdown on the tokio loop, and I observed it able to fix the `test_proc_mesh_size` finalization issues. However, I have to avoid initializing pyo3_async_runtime.

One way to make this shippable would be to remove our usein of pyo3_async_runtime entirely. We primarily need it for future_into_py but we can just reimplement it directly ourself.

Differential Revision: [D79533010](https://our.internmc.facebook.com/intern/diff/D79533010/)

[ghstack-poisoned]