[monarch] hyperactor: make sure that mock MAST jobs are unique
Without this, we cannot guarantee that the various world/proc names are unique.

This was not a problem before D79478197 because each mesh was private -- the underlying agents were not wired together.

However, since the mock MAST allocator uses the LocalAllocator to allocate host-local procs, they are now all wired together, and messages, and subsequently, we have overlapping routing entries.

Differential Revision: [D79596215](https://our.internmc.facebook.com/intern/diff/D79596215/)

ghstack-source-id: 300660703
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/756