[hyperactor] simplify supervision propagation; unhandled events always cause failure
Currently (local) supervision can propagate events without also killing an intermediate actor.

This is 1) wrong; and 2) complicated.

Instead, we treat an unhandled supervision event as an actor failure, and then reduce the propagation paths to one: that of an actor failing.

In order to retain accurate attribution, we add a "caused_by" field to the actor supervision events.

Differential Revision: [D79702385](https://our.internmc.facebook.com/intern/diff/D79702385/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D79702385/)!

[ghstack-poisoned]