Update base for Update on "[hyperactor] channel: allow ":" as a channel-type delimiter"
We currently use "!" to delimit between channel type and its address. The main issue with "!" is that it is not shell-friendly, as Bourne shells will expand "!" into a history entry.

With this change, we allow ":" to be used instead. With the previous change (requiring channel types to be specified), we can parse these two cases unambiguously.

The ChannelAddr `Display` implementation is adjusted to emit ":"-delimited channel addresses, too. It is the new preferred syntax.

Differential Revision: [D79846852](https://our.internmc.facebook.com/intern/diff/D79846852/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D79846852/)!

[ghstack-poisoned]