CanSend knows the sender, ActorMesh does not
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/808

Previously the CanSend capability did not know what actor was sending, whereas casting to an actor mesh required the sending actor.

This moves actor_id into CanSend since all of our implementations of CanSend easily already know the actor_id they are sending.

This change lets us take an explicit mailbox for sending from PythonActorMesh. This is important for de-futuring actor spawns because the mailbox we are sending from is now the client mailbox not the proc mesh mailbox. We have to be consistent about using it otherwise messages will go out of order.

This also refactors all the functionality out of the ActorMeshProtocol such as shape/proc_mesh which do not vary across the implementations.

ghstack-source-id: 302126566
@exported-using-ghexport

Differential Revision: [D79925866](https://our.internmc.facebook.com/intern/diff/D79925866/)