Update on "[monarch] How many controllers would a ControllerController control?"
This diff introduces the `get_or_spawn_controller(name, Class, *args, **kwargs)` API, which creates a singleton actor of type `Class` indexed by `name` if no such actor with `name` already exists; otherwise, it returns a reference to the existing singleton actor with that `name`. The singleton actor is created on the root client.

To facilitate this API and manage all of the singleton controllers, this diff adds the `_ControllerController` singleton, which lives on the root client and is spawned the first time a proc mesh is created. Every time a new actor mesh is created, the mesh receives a reference to the `_ControllerController`, which is added to `MonarchContext` and is accessed inside `get_or_spawn_controller`.

With the new `get_or_spawn_controller` functionality, we no longer need to initialize `_RdmaManager` and `DebugManager` (or any manager) inside `proc_mesh.py`. (Actually, `DebugManager` is no longer needed at all, and `DebugClient` has been renamed to `DebugController`).

In order to support the `_RdmaManager` pattern, where a controller singleton spawns a manager actor on a user-specified proc mesh, we need a way to pass proc meshes through actor endpoints. This diff implements the exceedingly hacky `ProcMeshRef`, which just contains the proc mesh id, and can't do anything `ProcMesh` can do. However, when passed to an actor endpoint on the root client process, `ProcMeshRef` can be dereferenced into an actual `ProcMesh` by looking up its id in the `_proc_mesh_registry`, which is a global map from proc mesh id to `weakref(ProcMesh)`.

Also, `rdma/__init__.pyi` was moved into `rdma.pyi` to resolve some pesky lint issues related to importing `monarch._rust_bindings.rdma`.

Differential Revision: [D80374416](https://our.internmc.facebook.com/intern/diff/D80374416/)

[ghstack-poisoned]