[hyperactor] net: zero copy framer
This change introduces a zero-copy framer, which will also be easily extendable to doing vectorized framing of multipart messages.

We eschew tokio's framer (which relies on queuing for cancellation safety), in favor of a simple(r) implementation: the reader maintains a simple state machine, while the writer requires the caller to maintain an explicit write state (since this has to be driven across selects in order to be made cancellation safe).

In this way, we have an easily hackable framer that does not introduce additional queues.

Differential Revision: [D80365228](https://our.internmc.facebook.com/intern/diff/D80365228/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D80365228/)!

[ghstack-poisoned]