[hyperactor] channel-level ping-pong benchmarks
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/906

This is an attempt to do an apples-to-apples comparison to P1903314366, to eliminate any non-channel related overheads.

The results replicate previous findings: that our throughput is hampered by excess data copies (either outright or through growing buffers in the encoding stack), and of tokio-level network i/o overheads. Both are being addressed. These benchmarks should help to serve as validation as this work lands.
ghstack-source-id: 304032339
@exported-using-ghexport

Differential Revision: [D80260732](https://our.internmc.facebook.com/intern/diff/D80260732/)