[monarch][debugger] Support breakpoints in code that was pickled by value
AKA "jank it into the linecache"

Functions in python have an attribute `__code__`, which itself has a property `co_filename`, which stores the name of the file where the function was defined. This is the primary way that `pdb` is able to display human-readable lines of code at each step of execution: from the current stack frame, it gets the associated filename and loads the source code using a module called `linecache`.

Now, monarch uses `cloudpickle` to serialize user-defined code so that it can run on remote hosts. In the simplest cases, `cloudpickle` serializes types *by reference*, which means the representation of `my_module.MyActorClass` is basically just `from my_module import MyActorClass`, and deserializing is just running the import. When the code runs on the remote host, it has `__code__.co_filename == /path/to/my_module/on/host`, because it was actually imported from wherever the code exists. Pdb is then very happy, because it knows how to display source code for debugging.

However, there are certain cases where `cloudpickle` either (a) assumes the type cannot be imported on the receiving end (e.g., when the type is defined in `__main__`), or (b) knows the type cannot be imported on the receiving end (e.g., if the object is a closure or a lambda). In these cases, `cloudpickle` serializes *by value*, which means it packages all of the relevant code, modules, globals, captured variables, etc. into a format such that the receiver can reconstruct the original object. As a consequence, when the code runs on the remote host, it has `__code__.co_filename == /path/on/user/not/remote`. Pdb is then very sad, because it doesn't know how to find the source code.

We can work around this issue by populating the `linecache` ourselves. Inside `PdbWrapper`, if we encounter a code path that doesn't exist on the remote host, we will load the source code for that path using the new `DebugController.get_source` endpoint and insert the code into `linecache` (this works because `DebugController` runs on the root client host,). Then, any time we encounter that path in the future, we can just retrieve the source from the local `linecache`.

To support this, `PdbWrapper` needs to check and (possibly) populate the `linecache` in each of these scenarios:
1. Before every `pdb` command.
2. When `set_trace` is called.
3. When `format_stack_entry` is called (`where`, `bt`, and exception tracebacks).

The monarch debugger should now be able to support breakpoints in actors defined in `__main__` as well as in closures

Differential Revision: [D80966586](https://our.internmc.facebook.com/intern/diff/D80966586/)

ghstack-source-id: 305608073
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/993