[hyperactor] channel example: make it into a flexible stress test
With this change, we can stress channels better by running the server and client independently (and potentially on separate machines). By default, it runs in "infinite" mode, reporting stats every 1000 messages.

The idea is that we can use this as a test bench to try to replicate issues we see in production with channels.

Differential Revision: [D81099840](https://our.internmc.facebook.com/intern/diff/D81099840/)

[ghstack-poisoned]