Update base for Update on "[monarch] view: make mapping more explicit, and make views extensible"
We try to better rationalize the various data structures in the `view` module, and attempt to reify their mapping in the API itself, so that concepts like extents and ranks have a consistent meaning.

Specifically, we:

- Introduce the concept of a `Region`. A region describes how to index a hyperrect into a (possibly) larger space.
- A `View` is now a collection indexed by a `Region`. The intent is to use `View`s as a trait for the various meshes and mesh references, to provide a consisten API to sub-slice them, etc. Views behave as much as possible like an ordinary Rust collection.

With this we implement `View` for `Region`. The "Item" of this implementation is the underlying physical rank. Thus the meaning of this collection is to map one space to another. This makes the changes mostly API compatible with the existing "View" (except that it is now a trait). However, when we now add other implementations (e.g. a HostMesh), the user is presented with a consistent definition: it is a collection of hosts, arranged into an extent.

In this case, if the user wishes to recover the underlying physical ranks, this can be done by calling `host_mesh.region().iter()`, so that the meaning of the various ranks at play are clear and consistent.

Here's the new module-level documentation:

```
//! The purpose of this module is to provide data structures to efficiently
//! describe subsets of nonnegative integers, called *ranks*, represented by
//! `usize` values. Ranks are organized into multi-dimensional spaces in which
//! each discrete point is a rank, mapped in row-major order.
//!
//! These dimensions represent a space that carry semantic meaning, such that
//! ranks are usually sub-set along some dimension of the space. For example,
//! ranks may be organized into a space with dimensions "replica", "host", "gpu",
//! and we'd expect to subset along these dimensions, for example to select all
//! GPUs in a given replica.
//!
//! This alignment helps provide a simple and efficient representation, internally
//! in the form of [`crate::Slice`], comprising an offset, sizes, and strides that
//! index into the space.
//!
//! - [`Extent`]: the *shape* of the space, naming each dimension and specifying
//!               their sizes.
//! - [`Point`]: a specific coordinate in an extent, together with its linearized rank.
//! - [`Region`]: a (possibly sparse) hyper-rectangle of ranks within a larger extent.
//!               Since it is always rectangular, it also defines its own extent.
//! - [`View`]: a collection of items indexed by [`Region`]. Views provide standard
//!             manipulation operations and use ranks as an efficient indexing scheme.
```

Differential Revision: [D81002285](https://our.internmc.facebook.com/intern/diff/D81002285/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D81002285/)!

[ghstack-poisoned]