[monarch, quick rfc] Less surprising actor lifetimes
This establishes that the lifetime of an actor or proc mesh is initially bound to the lifetime of the actor that spawned it.

In particular, an actor mesh or proc mesh will stop running when (1) its parent stops, or (2) its parent explicitly calls stop on it.

This replaces the previous surprising behavior where an actor_mesh/proc_mesh would stop once there was no remaining external reference to it.  This led to errors where renaming dimensions, slicing, or creating a proc with only one actor on it caused the thing to stop immediately.

Differential Revision: [D81141365](https://our.internmc.facebook.com/intern/diff/D81141365/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D81141365/)!

ghstack-source-id: 305984952
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1006