[hyperactor] channel example: make it into a flexible stress test
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1002

With this change, we can stress channels better by running the server and client independently (and potentially on separate machines). By default, it runs in "infinite" mode, reporting stats every 1000 messages.

The idea is that we can use this as a test bench to try to replicate issues we see in production with channels.
ghstack-source-id: 305992192
@exported-using-ghexport

Differential Revision: [D81099840](https://our.internmc.facebook.com/intern/diff/D81099840/)