Update on "[monarch][debugger] Support breakpoints in code that was pickled by value"
When `pdb` wants to get the human-readable source code at each step of execution, its primary way of doing so boils down to calling into the `linecache` module. `linecache` attempts to load the file from disk, or if the file doesn't exist, it looks for the `__loader__` attribute in its `module_globals` argument. The loader needs to provide a `get_source` method that returns the relevant source code. In the context of `pdb`, when checking the `linecache` for a given function's source, the filename is `<the_function>.__code__.co_filename`, and the value of `module_globals` is `<the_function>.__globals__['__loader__']`.

Now, monarch uses `cloudpickle` to serialize user-defined code so that it can run on remote hosts. In the simplest cases, `cloudpickle` serializes types *by reference*, which means the representation of `my_module.MyActorClass` is basically just `from my_module import MyActorClass`, and deserializing is just running the import. When the code runs on the remote host, it has `__code__.co_filename == /path/to/my_module/on/host`, because it was actually imported from wherever the code exists. Pdb is then very happy, because it knows how to display source code for debugging.

However, there are certain cases where `cloudpickle` either (a) assumes the type cannot be imported on the receiving end (e.g., when the type is defined in `__main__`), or (b) knows the type cannot be imported on the receiving end (e.g., if the object is a closure or a lambda). In these cases, `cloudpickle` serializes *by value*, which means it packages all of the relevant code, modules, globals, captured variables, etc. into a format such that the receiver can reconstruct the original object. As a consequence, when the code runs on the remote host, it has `__code__.co_filename == /path/on/user/not/remote`. Pdb is then very sad, because it doesn't know how to find the source code.

This diff works around this issue in two ways:
1. We hook into the unpickler and detect when a function was pickled by value. When this happens, we set `<the_function>.__globals__['__loader__']` to an instance of the new `RemoteImportLoader` class, which loads the source code for `<the_function>.__code__.co_filename` from the root client host.
2. There are a few places that `pdb` calls into `linecache` just based on the filename, without any module globals. To account for this, we can override `pdb.Pdb.lookupmodule` to load the source code from the root client host if the relevant file isn't found.

The monarch debugger should now be able to support breakpoints in actors defined in `__main__` as well as in closures.

Differential Revision: [D80966586](https://our.internmc.facebook.com/intern/diff/D80966586/)

[ghstack-poisoned]