fix up demo
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1013

create a version of demo.py that uses the new apis (proc_mesh/hosts no device mesh). Remove the parts of the demo (pipes, wandb, etc.) that are not directly related to the tensor engine.

Do some fixup to make it work again, including some slightly more sophisticated logic to decide what dimension is the gpu dimension in a proc_mesh.
ghstack-source-id: 306661337
@exported-using-ghexport

Differential Revision: [D81178894](https://our.internmc.facebook.com/intern/diff/D81178894/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D81178894/)!