Example scheme to use for value mesh compression
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1181

We will eventually implement value mesh compression where we use tree-based aggregation to ensure we have O(log n) latency to respond to a call everything is returning the same value or there is only O(1) values that are not the same value. This is primarily to handle things like a barrier where everything is returning None and we want the response to be efficient.

This scheme uses a run-line encoding of common entries for adjacent ranks. It stores a sorted list of (start, end, bytes) and metadata to be able to answer whether we have a dense list of entries. Entries can be merged, coaescing start/end ranges that accumulate.

Differential Revision: [D82240319](https://our.internmc.facebook.com/intern/diff/D82240319/)
ghstack-source-id: 309156347